Move routes to the same screen file
