+++ app/src/androidTest/java/com/marcopla/flashcards/presentation/edit/EditScreenRobot.kt
@@ -16,8 +16,8 @@ import com.marcopla.flashcards.data.repository.FlashCardRepositoryImpl
 import com.marcopla.flashcards.domain.usecase.DeleteUseCase
 import com.marcopla.flashcards.domain.usecase.EditUseCase
 import com.marcopla.flashcards.domain.usecase.LoadUseCase
-import com.marcopla.flashcards.presentation.navigation.EditRoute
 import com.marcopla.flashcards.presentation.navigation.FLASH_CARD_ID_ARG_KEY
+import com.marcopla.flashcards.presentation.screen.edit.EditRoute
 import com.marcopla.flashcards.presentation.screen.edit.EditViewModel
 import com.marcopla.testing_shared.FakeFlashCardDao
 import com.marcopla.testing_shared.TestFlashCardRepository
