+++ app/src/main/java/com/marcopla/flashcards/presentation/navigation/AppNavHost.kt
@@ -1,25 +1,15 @@
 package com.marcopla.flashcards.presentation.navigation
 
-import androidx.compose.material.rememberScaffoldState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import com.marcopla.flashcards.presentation.screen.add.AddScreen
-import com.marcopla.flashcards.presentation.screen.add.AddViewModel
-import com.marcopla.flashcards.presentation.screen.add.HandleInfoTextEffect
+import com.marcopla.flashcards.presentation.screen.add.AddRoute
 import com.marcopla.flashcards.presentation.screen.carousel.CarouselScreen
-import com.marcopla.flashcards.presentation.screen.edit.DeleteConfirmationDialog
-import com.marcopla.flashcards.presentation.screen.edit.EditScreen
-import com.marcopla.flashcards.presentation.screen.edit.EditViewModel
-import com.marcopla.flashcards.presentation.screen.home.HomeScreen
-import com.marcopla.flashcards.presentation.screen.home.HomeScreenState
-import com.marcopla.flashcards.presentation.screen.home.HomeViewModel
+import com.marcopla.flashcards.presentation.screen.edit.EditRoute
+import com.marcopla.flashcards.presentation.screen.home.HomeRoute
 import com.marcopla.flashcards.presentation.screen.result.ResultsScreen
 
 @Composable
@@ -75,83 +65,6 @@ fun AppNavHost(
             )
         }
     }
-}
-
-@Composable
-private fun HomeRoute(
-    onNavigateToAddScreen: () -> Unit,
-    onItemClicked: (Int) -> Unit,
-    onNavigateToCarouselScreen: () -> Unit,
-    viewModel: HomeViewModel = hiltViewModel()
-) {
-    val screenState: HomeScreenState by viewModel.homeScreenState.collectAsStateWithLifecycle()
-    HomeScreen(
-        screenState = screenState,
-        onNavigateToAddScreen = onNavigateToAddScreen,
-        onItemClicked = onItemClicked,
-        onNavigateToCarouselScreen = onNavigateToCarouselScreen
-    )
-}
-
-@Composable
-fun AddRoute(viewModel: AddViewModel = hiltViewModel()) {
-    val infoTextState = viewModel.infoTextState.value
-    val scaffoldState = rememberScaffoldState()
-    HandleInfoTextEffect(
-        infoTextState.messageStringRes,
-        scaffoldState.snackbarHostState
-    )
-
-    AddScreen(
-        frontTextState = viewModel.frontTextState.value,
-        backTextState = viewModel.backTextState.value,
-        addScreenState = viewModel.addScreenState.value,
-        onSubmitClick = {
-            viewModel.attemptSubmit(
-                frontText = viewModel.frontTextState.value.text,
-                backText = viewModel.backTextState.value.text
-            )
-        },
-        onFrontTextValueChange = { frontInput -> viewModel.updateFrontText(frontInput) },
-        onBackTextValueChange = { backInput -> viewModel.updateBackText(backInput) },
-        scaffoldState = scaffoldState
-    )
-}
-
-@Composable
-fun EditRoute(
-    viewModel: EditViewModel = hiltViewModel(),
-    onPopBackStack: () -> Unit
-) {
-    if (viewModel.shouldShowDeleteConfirmation.value) {
-        DeleteConfirmationDialog(
-            onConfirmationClick = viewModel::delete,
-            onCancelClick = viewModel::hideDeleteConfirmationDialog,
-            onDismissRequest = viewModel::hideDeleteConfirmationDialog
-        )
-    }
-
-    EditScreen(
-        onFlashCardEdited = onPopBackStack,
-        onFlashCardDeleted = {
-            viewModel.hideDeleteConfirmationDialog()
-            onPopBackStack()
-        },
-        editScreenState = viewModel.screenState.value,
-        onReset = viewModel::reset,
-        onShowDeleteConfirmationDialog = viewModel::showDeleteConfirmationDialog,
-        onSubmit = {
-            viewModel.attemptSubmit(
-                viewModel.frontTextState.value.text,
-                viewModel.backTextState.value.text
-            )
-        },
-        onInitState = viewModel::initState,
-        frontTextState = viewModel.frontTextState.value,
-        onFrontTextValueChange = viewModel::updateFrontText,
-        backTextState = viewModel.backTextState.value,
-        onBackTextValueChange = viewModel::updateBackText
-    )
 }
 
 object Routes {
