+++ app/src/main/java/com/marcopla/flashcards/presentation/screen/home/HomeScreen.kt
@@ -8,6 +8,7 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.PlayArrow
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -16,10 +17,28 @@ import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.marcopla.flashcards.R
 import com.marcopla.flashcards.data.model.FlashCard
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
+
+@Composable
+fun HomeRoute(
+    onNavigateToAddScreen: () -> Unit,
+    onItemClicked: (Int) -> Unit,
+    onNavigateToCarouselScreen: () -> Unit,
+    viewModel: HomeViewModel = hiltViewModel()
+) {
+    val screenState: HomeScreenState by viewModel.homeScreenState.collectAsStateWithLifecycle()
+    HomeScreen(
+        screenState = screenState,
+        onNavigateToAddScreen = onNavigateToAddScreen,
+        onItemClicked = onItemClicked,
+        onNavigateToCarouselScreen = onNavigateToCarouselScreen
+    )
+}
 
 @Composable
 fun HomeScreen(
